Moves the middleware defined in the identity service
All middleware really belongs in the middleware library. Also, moving it
removed the hard dependency of the catalog on the identity service.

	modified:   lib/middleware/index.js
	new file:   lib/middleware/request-authenticate.js
	new file:   lib/middleware/request-authorize.js
	modified:   lib/services/catalog/index.js
	modified:   lib/services/identity/index.js
