+++ lib/middleware/index.js
@@ -15,6 +15,8 @@ const MIDDLEWARE = {
 	},
 	'error-handler': require('./error-handler'),
 	'request-accept': require('./request-accept'),
+	'request-authenticate': require('./request-authenticate'),
+	'request-authorize': require('./request-authorize'),
 	'request-json-api': require('./request-json-api'),
 	'request-options': require('./request-options'),
 	'response-cache-control': require('./response-cache-control'),
