+++ lib/services/catalog/index.js
@@ -3,6 +3,7 @@
 const _ = require('lodash');
 const express = require('express');
 
+const middleware = require('../../middleware');
 const initializeQueries = require('./queries/');
 const initializeCommands = require('./commands/');
 const CatalogItemController = require('./controllers/catalog-item-controller');
@@ -119,7 +120,7 @@ module.exports = function (bus, options) {
 		types.forEach(type => {
 			router.all(
 				`/${type}s`,
-				bus.query({middleware: 'authorize'}, {bus, audience: {
+				middleware['request-authorize']({bus, audience: {
 					get: ['admin', 'platform'],
 					post: ['admin'],
 					delete: ['admin']
@@ -129,7 +130,7 @@ module.exports = function (bus, options) {
 
 			router.all(
 				`/${type}s/:id`,
-				bus.query({middleware: 'authorize'}, {bus, audience: {
+				middleware['request-authorize']({bus, audience: {
 					get: ['admin', 'platform'],
 					patch: ['admin'],
 					delete: ['admin']
@@ -138,12 +139,18 @@ module.exports = function (bus, options) {
 			);
 		});
 
-		router.all('/search', CatalogSearchController.create({bus}));
+		router.all(
+			'/search',
+			middleware['request-authorize']({bus, audience: {
+				get: ['platform']
+			}}),
+			CatalogSearchController.create({bus})
+		);
 
 		specTypes.forEach(type => {
 			router.all(
 				`/${type}s`,
-				bus.query({middleware: 'authorize'}, {bus, audience: {
+				middleware['request-authorize']({bus, audience: {
 					get: ['admin'],
 					post: ['admin'],
 					delete: ['admin']
@@ -153,7 +160,7 @@ module.exports = function (bus, options) {
 
 			router.all(
 				`/${type}s/:id`,
-				bus.query({middleware: 'authorize'}, {bus, audience: {
+				middleware['request-authorize']({bus, audience: {
 					get: ['admin'],
 					patch: ['admin'],
 					delete: ['admin']
