+++ lib/services/identity/index.js
@@ -2,14 +2,15 @@
 
 const _ = require('lodash');
 const express = require('express');
-const Boom = require('boom');
-const debug = require('debug');
+const debug = require('debug')('oddworks:identity-service');
+const middleware = require('../../middleware');
 const initializeQueries = require('./queries/');
 const IdentityItemController = require('./controllers/identity-item-controller');
 const IdentityListController = require('./controllers/identity-list-controller');
 const IdentityConfigController = require('./controllers/identity-config-controller');
 
 module.exports = function (bus, options) {
+	debug('initializing');
 	if (!bus || !_.isObject(bus)) {
 		throw new Error('The bus must be the first argument.');
 	}
@@ -53,108 +54,8 @@ module.exports = function (bus, options) {
 		queries.composeConfig
 	);
 
-	service.middleware = {
-		// options.header - String default=Authorization
-		authenticate(options) {
-			const log = debug('oddworks:middleware:authenticate');
-			options = _.defaults({}, options, {
-				header: 'Authorization'
-			});
-
-			const header = options.header;
-			log('using the %s header', header);
-
-			return function authenticateMiddleware(req, res, next) {
-				// Parse out the Authorization header if used.
-				// ex: "Authorization: Bearer eyJ0eXAiOiJKV1QiLCJh..."
-				const headerValue = (req.get(header) || '').split(/[\s]+/);
-				const token = headerValue.length > 1 ? headerValue[1] : headerValue[0];
-
-				if (token) {
-					log('using token %s', token);
-					bus
-						.query({role: 'identity', cmd: 'verify'}, {token})
-						.then(identity => {
-							req.identity = identity;
-							next();
-							return null;
-						})
-						.catch(err => {
-							log('invalid-access-token: %s', err.message);
-							bus.broadcast(
-								{level: 'warn', event: 'invalid-access-token'},
-								{message: err.message, error: err}
-							);
-							next(Boom.unauthorized('Invalid Access Token'));
-							return null;
-						});
-				} else {
-					log('missing-access-token');
-					bus.broadcast(
-						{level: 'info', event: 'missing-access-token'},
-						{message: 'Missing Access Token'}
-					);
-					next(Boom.unauthorized('Missing Access Token'));
-				}
-			};
-		},
-
-		// options.audience - Object
-		authorize(options) {
-			const log = debug('oddworks:middleware:authorize');
-			options = _.defaults({}, options, {
-				audience: Object.create(null)
-			});
-
-			const whiteList = options.audience;
-
-			return function authorizeMiddleware(req, res, next) {
-				let audience = (req.identity || {}).audience;
-				if (!audience) {
-					log('invalid-access-token: Missing token audience claim.');
-					bus.broadcast(
-						{level: 'warn', event: 'invalid-access-token'},
-						{message: 'Missing token audience claim'}
-					);
-					next(Boom.unauthorized('Missing Token audience claim'));
-				}
-
-				audience = Array.isArray(audience) ? audience : [audience];
-
-				const method = req.method.toLowerCase();
-				const allowed = whiteList[method];
-				log('method: %s', method);
-				log('audience: %s', audience.join());
-				log('allowed: %s', allowed.join());
-				if (!allowed) {
-					return next(Boom.methodNotAllowed());
-				}
-
-				let i;
-				for (i = 0; i < allowed.length; i += 1) {
-					if (audience.indexOf(allowed[i]) >= 0) {
-						return next();
-					}
-				}
-
-				next(Boom.forbidden(
-					`${req.method} access not permitted to this resource`
-				));
-			};
-		}
-	};
-
-	bus.queryHandler(
-		{middleware: 'authenticate'},
-		service.middleware.authenticate
-	);
-
-	bus.queryHandler(
-		{middleware: 'authorize'},
-		service.middleware.authorize
-	);
-
 	service.router = function (options) {
+		debug('setting up routes');
 		options = _.defaults({}, options, {
 			types: ['channel', 'platform']
 		});
@@ -165,7 +66,7 @@ module.exports = function (bus, options) {
 		types.forEach(type => {
 			router.all(
 				`/${type}s`,
-				service.middleware.authorize({audience: {
+				middleware['request-authorize']({audience: {
 					get: ['admin'],
 					post: ['admin']
 				}}),
@@ -174,7 +75,7 @@ module.exports = function (bus, options) {
 
 			router.all(
 				`/${type}s/:id`,
-				service.middleware.authorize({audience: {
+				middleware['request-authorize']({audience: {
 					get: ['admin'],
 					patch: ['admin'],
 					delete: ['admin']
@@ -185,7 +86,7 @@ module.exports = function (bus, options) {
 
 		router.all(
 			'/config',
-			service.middleware.authorize({audience: {
+			middleware['request-authorize']({audience: {
 				get: ['platform', 'admin']
 			}}),
 			IdentityConfigController.create({bus})
@@ -194,5 +95,6 @@ module.exports = function (bus, options) {
 		return router;
 	};
 
+	debug('initialized');
 	return Promise.resolve(service);
 };
